Moved GenericOpenRewriteRecipe to sbm-support-rewrite
